display: move `Rect` into `rect.cpp/.h`
